requirements.txt
JACK-Client
